update crf entity extractor
